refactor(HttpResponseValues): use http request to extract URI
