+++ src/main/java/org/opentripplanner/middleware/utils/HttpResponseHandler.java
@@ -2,38 +2,26 @@ package org.opentripplanner.middleware.utils;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.client.ResponseHandler;
-import org.apache.http.util.EntityUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
+import org.apache.http.client.methods.HttpUriRequest;
 
 /**
  * Extracts the required values from an {@link HttpResponse} and populates an instance of {@link HttpResponseValues}.
  */
 public class HttpResponseHandler implements ResponseHandler<HttpResponseValues> {
-    private static final Logger LOG = LoggerFactory.getLogger(HttpResponseHandler.class);
+    private final HttpUriRequest request;
+
+    /**
+     * Construct the handler with the incoming {@link HttpUriRequest} for access to request fields if needed.
+     */
+    public HttpResponseHandler(HttpUriRequest request) {
+        this.request = request;
+    }
 
     /**
      * Populates the {@link HttpResponseValues} with required values. Once this method completes the http connection
      * is closed.
      */
     public HttpResponseValues handleResponse(final HttpResponse response) {
-        return new HttpResponseValues(response, getResponseBodyAsString(response));
+        return new HttpResponseValues(request, response);
     }
-
-    /**
-     * Extract the response body as a String. If an exception occurs return null.
-     */
-    private static String getResponseBodyAsString(HttpResponse response) {
-        String responseBody = null;
-        try {
-            responseBody = EntityUtils.toString(response.getEntity());
-            EntityUtils.consume(response.getEntity());
-        } catch (IOException e) {
-            LOG.error("Unable to get response body", e);
-        }
-        return responseBody;
-    }
-
 }
